+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.test.ts
@@ -143,9 +143,9 @@ describe('QueryStepsDataSource', () => {
     });
 
     it('should throw too many requests error when API returns 429 status', async () => {
-      backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' }))
-        .mockReturnValue(createFetchError(429));
+      jest.spyOn(datastore, 'post').mockImplementation(() => {
+        throw new Error('Request failed with status code: 429');
+      });
 
       await expect(datastore.querySteps()).rejects.toThrow(
         'The query to fetch steps failed due to too many requests. Please try again later.'